Initial version of WAT space invaders clone
This is the initial upload of the WAT space invaders clone.  It has a lot of the basics.  I used a tool I created called wat-wasm that is available on npm.  It allows #include WAT tags so that I could break up the WAT into multiple files instead of having it all in one large file.